Update dm_control_compatibility.py
There seems to be a bug at present, due to the fact that dm_control environments re-compile their physics (including the Model and Data structures) on each reset (to allow for, e.g., domain randomization). As such, rendering stops working after the first reset().
This commit fixes the problem by re-creating a new Viewer in the "human" mode.

Note that this problem is only related to "human" mode rendering, since the other pixel-based renderings call self._env.physics directly instead, which is guaranteed to be the latest (correct) object.